+++ Pyro/src/pyro/graphics/raw_model.h
@@ -6,8 +6,8 @@ namespace pyro
     {
     public:
         raw_model() = default;
-        raw_model(unsigned int vao_id, int vertex_count);
-        void create(unsigned int vao_id, int vertex_count);
+        raw_model(unsigned int vao_id, size_t vertex_count);
+        void create(unsigned int vao_id, size_t vertex_count);
 
         unsigned int vao_id() const;
         int vertex_count() const;
